refactor: modularize monolithic CLI tools
- Extracted streamCTRL_TUI logic into input and render modules.
- Split overview_data into sys, cache, scan, graph, and sort modules.
- Modularized milk-perfbench into setup, perf, sysmon, and report.
- Updated CMakeLists.txt to build new targets properly.
