+++ src/engine/libprocessinfo/procCTRL/procCTRL_TUI.c
@@ -158,382 +158,247 @@ static inline void *link_scan_shm(const char *name, size_t size) {
 
 #include <sys/resource.h>
 
-errno_t processinfo_CTRLscreen()
-{
-    if (getenv("PROCCTRL_DEBUG")) procCTRL_debug_mode = 1;
-    if (procCTRL_debug_mode) printf("DEBUG: processinfo_CTRLscreen start\n");
-
-    FILE *flog = NULL;
-    if (strlen(procCTRL_logfile) > 0) {
-        flog = fopen(procCTRL_logfile, "a");
-        if (flog) {
-            fprintf(flog, "\n--- processinfo_CTRLscreen started ---\n");
-            fflush(flog);
-        }
-    }
-
-    if (flog) { fprintf(flog, "Checking for daemon...\n"); fflush(flog); }
-    if (system("pgrep \"milk-procCTRL-s\" > /dev/null") != 0) {
-        fprintf(stderr, "\nWARNING: milk-procCTRL-scan daemon is not running.\n");
-        printf("Start it now in tmux session 'milk-procCTRL-scan'? [y/n] ");
-        fflush(stdout);
-        char response = 'n';
-        if (scanf(" %c", &response) == 1 && (response == 'y' || response == 'Y')) {
-            printf("Launching milk-procCTRL-scan...\n");
-            if(system("tmux new-session -d -s milk-procCTRL-scan 'milk-procCTRL-scan'") < 0) {}
-            sleep(1);
-            if (system("pgrep \"milk-procCTRL-s\" > /dev/null") != 0) {
-                fprintf(stderr, "ERROR: Failed to launch milk-procCTRL-scan daemon.\n");
-                if (flog) fclose(flog);
-                return RETURN_FAILURE;
-            }
-        } else {
-            fprintf(stderr, "ERROR: milk-procCTRL-scan daemon is required for this tool.\n");
-            if (flog) fclose(flog);
-            return RETURN_FAILURE;
-        }
-    }
-
-    processinfo_procdirname(local_shmdir);
+typedef struct {
+    PROCINFOPROC *procinfoproc;
+    PROCSCAN_SHM *scan_shm;
+    FILE *flog;
     char procdname[STRINGMAXLEN_DIRNAME];
-    processinfo_procdirname(procdname);
-
-    if (flog) { fprintf(flog, "Allocating procinfoproc...\n"); fflush(flog); }
-    PROCINFOPROC *procinfoproc = (PROCINFOPROC *) calloc(1, sizeof(PROCINFOPROC));
-    if(procinfoproc == NULL) {
-        PRINT_ERROR("calloc returns NULL pointer");
-        if (flog) fclose(flog);
-        return RETURN_FAILURE;
-    }
-
-    procinfoproc->NBcpus = GetNumberCPUs(procinfoproc);
-    GetCPUloads(procinfoproc); 
-
-    if(system("which cset > /dev/null 2>&1") == 0) procinfoproc->has_cset = 1;
-    else procinfoproc->has_cset = 0;
-
-    // Initialize column visibility and sorting
-    for(int m=0; m<10; m++) {
-        for(int i=0; i<10; i++) procinfoproc->col_visible[m][i] = 1;
-        procinfoproc->sort_col[m] = 0;
-        procinfoproc->sort_dir[m] = 0;
-        procinfoproc->sort_mode[m] = m;
-    }
-    procinfoproc->selected_col = 1;
-
-    STRINGLISTENTRY *CPUsetList = (STRINGLISTENTRY *) malloc(sizeof(STRINGLISTENTRY) * 1000);
-    int NBCPUset __attribute__((unused)) = processinfo_CPUsets_List(CPUsetList, procinfoproc->has_cset);
-
-    if (flog) { fprintf(flog, "Connecting to process list...\n"); fflush(flog); }
-    if(processinfo_shm_list_create() == -1) {
-        printf("==== ERROR: CANNOT ACCESS PROCESS LIST ====\n");
-        if (flog) fclose(flog);
-        return RETURN_FAILURE;
-    }
-    procinfoproc->pinfolist = pinfolist;
-    
-    char scan_shm_name[STRINGMAXLEN_FULLFILENAME];
-    snprintf(scan_shm_name, sizeof(scan_shm_name), "%s/%s", procdname, PROCESSINFO_SCAN_SHM_NAME);
-    if (flog) { fprintf(flog, "Linking to scan SHM: %s\n", scan_shm_name); fflush(flog); }
-    PROCSCAN_SHM *scan_shm = (PROCSCAN_SHM *) link_scan_shm(scan_shm_name, sizeof(PROCSCAN_SHM));
-    if (scan_shm == MAP_FAILED) {
-        printf("WARNING: Could not link to scan SHM %s. Stats may be missing.\n", scan_shm_name);
-        scan_shm = NULL;
-    }
-
-    TUI_set_screenprintmode(SCREENPRINT_NCURSES);
-    if(getenv("MILK_TUIPRINT_STDIO")) TUI_set_screenprintmode(SCREENPRINT_STDIO);
-    
-    if (flog) { fprintf(flog, "Initializing terminal...\n"); fflush(flog); } // Corrected 'frow' to 'flog'
-    ansi_raw_mode_enter();
-    TUI_init_terminal(&wrow, &wcol);
-    if(wrow < 10) wrow = 10;
-
-    if (flog) { fprintf(flog, "Allocating pinfodisp buffer (250MB)\n"); fflush(flog); }
-    procinfoproc->pinfodisp = (PROCESSINFODISP *) calloc(PROCESSINFOLISTSIZE, sizeof(PROCESSINFODISP));
-    if (procinfoproc->pinfodisp == NULL) {
-        ansi_raw_mode_exit();
-        fprintf(stderr, "FATAL ERROR: Could not allocate 250MB process info buffer.\n");
-        if (flog) fclose(flog);
-        return RETURN_FAILURE;
-    }
-    procinfoproc->NBpinfodisp = PROCESSINFOLISTSIZE;
-    
-    procinfoproc->DisplayMode = PROCCTRL_DISPLAYMODE_CTRL;
-    procinfoproc->loop = 1;
-    procinfoproc->twaitus = 1000000;
-
-    int backstderr = -1, newstderr = -1;
-    if (procCTRL_debug_mode == 0) {
-        fflush(stderr);
-        backstderr = dup(STDERR_FILENO);
-        newstderr  = open("/dev/null", O_WRONLY);
-        dup2(newstderr, STDERR_FILENO);
-        close(newstderr);
-    }
-
-    struct timespec t_last_scan = {0,0}, t_now;
-    float frequ = 32.0;
-    int pindexSelected = -1;
-    int pindexActiveSelected = 0;
-    long doffsetindex = 0;
-    int freeze = 0;
-    int loopOK = 1;
-    int Xexit __attribute__((unused)) = 0;
+    float frequ;
+    int pindexSelected;
+    int pindexActiveSelected;
+    long doffsetindex;
+    int freeze;
+    int loopOK;
+    int Xexit;
     char monstring[200];
-    int monstringlen = 200;
-    int last_ch = -1;
+    int monstringlen;
+    int last_ch;
+    float tool_cpu_pcnt;
+    float actual_fps;
+    
+    struct timespec t_last_scan;
+    struct timespec t_now;
+    struct rusage usage_prev;
+    struct rusage usage_cur;
+    struct timespec t_usage_prev;
+    struct timespec t_usage_cur;
+    struct timespec t_disp_prev;
+    struct timespec t_disp_cur;
+} procctrl_context_t;
 
-    struct rusage usage_prev, usage_cur;
-    struct timespec t_usage_prev, t_usage_cur;
-    getrusage(RUSAGE_SELF, &usage_prev);
-    clock_gettime(CLOCK_MONOTONIC, &t_usage_prev);
-    float tool_cpu_pcnt = 0.0;
+static void procctrl_render_frame(procctrl_context_t *ctx, int NBactive);
 
-    struct timespec t_disp_prev, t_disp_cur;
-    clock_gettime(CLOCK_MONOTONIC, &t_disp_prev);
-    float actual_fps = 0.0;
-
-    if (flog) { fprintf(flog, "Entering main loop.\n"); fflush(flog); }
-    sc_frame_clear();
-
-    while(loopOK) {
-        if(processinfo_signal_SEGV) { 
-            if (flog) { fprintf(flog, "SEGV signal received!\n"); fflush(flog); }
-            loopOK=0; break; 
-        }
-
-        clock_gettime(CLOCK_MONOTONIC, &t_now);
-        double elapsed = (t_now.tv_sec - t_last_scan.tv_sec) + (t_now.tv_nsec - t_last_scan.tv_nsec) * 1e-9;
-        if (elapsed >= (procinfoproc->twaitus * 1e-6)) {
-            processinfo_scan_step(procinfoproc);
-            t_last_scan = t_now;
-            
-            getrusage(RUSAGE_SELF, &usage_cur);
-            clock_gettime(CLOCK_MONOTONIC, &t_usage_cur);
-            double t_diff = (t_usage_cur.tv_sec - t_usage_prev.tv_sec) + (t_usage_cur.tv_nsec - t_usage_prev.tv_nsec) * 1e-9;
-            double u_diff = (usage_cur.ru_utime.tv_sec - usage_prev.ru_utime.tv_sec) + (usage_cur.ru_utime.tv_usec - usage_prev.ru_utime.tv_usec) * 1e-6;
-            double s_diff = (usage_cur.ru_stime.tv_sec - usage_prev.ru_stime.tv_sec) + (usage_cur.ru_stime.tv_usec - usage_prev.ru_stime.tv_usec) * 1e-6;
-            if (t_diff > 0) tool_cpu_pcnt = 100.0 * (u_diff + s_diff) / t_diff;
-            usage_prev = usage_cur;
-            t_usage_prev = t_usage_cur;
-        }
-
-        clock_gettime(CLOCK_MONOTONIC, &t_disp_cur);
-        double d_elapsed = (t_disp_cur.tv_sec - t_disp_prev.tv_sec) + (t_disp_cur.tv_nsec - t_disp_prev.tv_nsec) * 1e-9;
-        if (d_elapsed > 0) actual_fps = 0.9 * actual_fps + 0.1 * (1.0 / d_elapsed);
-        t_disp_prev = t_disp_cur;
-
-        usleep((long)(1000000.0 / frequ));
-        int ch = get_singlechar_nonblock();
+static void procctrl_update_stats(procctrl_context_t *ctx) {
+    clock_gettime(CLOCK_MONOTONIC, &ctx->t_now);
+    double elapsed = (ctx->t_now.tv_sec - ctx->t_last_scan.tv_sec) + (ctx->t_now.tv_nsec - ctx->t_last_scan.tv_nsec) * 1e-9;
+    if (elapsed >= (ctx->procinfoproc->twaitus * 1e-6)) {
+        processinfo_scan_step(ctx->procinfoproc);
+        ctx->t_last_scan = ctx->t_now;
         
-        int NBactive = (scan_shm) ? scan_shm->NBactive : 0;
-        int m = procinfoproc->DisplayMode;
-        
-        // Sorting logic
-        if (procinfoproc->sort_col[m] > 0 && scan_shm != NULL) {
-            sort_ctx_m = procinfoproc->sort_mode[m];
-            sort_ctx_col = procinfoproc->sort_col[m];
-            sort_ctx_dir = procinfoproc->sort_dir[m];
-            sort_ctx_scan_shm = scan_shm;
-            sort_ctx_pinfolist = pinfolist;
-            
-            for(int i=0; i<NBactive; i++) procinfoproc->local_sorted_pindex[i] = scan_shm->sorted_pindex[i];
-            qsort(procinfoproc->local_sorted_pindex, NBactive, sizeof(int), proc_comp);
+        getrusage(RUSAGE_SELF, &ctx->usage_cur);
+        clock_gettime(CLOCK_MONOTONIC, &ctx->t_usage_cur);
+        double t_diff = (ctx->t_usage_cur.tv_sec - ctx->t_usage_prev.tv_sec) + (ctx->t_usage_cur.tv_nsec - ctx->t_usage_prev.tv_nsec) * 1e-9;
+        double u_diff = (ctx->usage_cur.ru_utime.tv_sec - ctx->usage_prev.ru_utime.tv_sec) + (ctx->usage_cur.ru_utime.tv_usec - ctx->usage_prev.ru_utime.tv_usec) * 1e-6;
+        double s_diff = (ctx->usage_cur.ru_stime.tv_sec - ctx->usage_prev.ru_stime.tv_sec) + (ctx->usage_cur.ru_stime.tv_usec - ctx->usage_prev.ru_stime.tv_usec) * 1e-6;
+        if (t_diff > 0) ctx->tool_cpu_pcnt = 100.0 * (u_diff + s_diff) / t_diff;
+        ctx->usage_prev = ctx->usage_cur;
+        ctx->t_usage_prev = ctx->t_usage_cur;
+    }
+
+    clock_gettime(CLOCK_MONOTONIC, &ctx->t_disp_cur);
+    double d_elapsed = (ctx->t_disp_cur.tv_sec - ctx->t_disp_prev.tv_sec) + (ctx->t_disp_cur.tv_nsec - ctx->t_disp_prev.tv_nsec) * 1e-9;
+    if (d_elapsed > 0) ctx->actual_fps = 0.9 * ctx->actual_fps + 0.1 * (1.0 / d_elapsed);
+    ctx->t_disp_prev = ctx->t_disp_cur;
+}
+
+static void procctrl_handle_keyboard_event(procctrl_context_t *ctx, int ch, int NBactive) {
+    ctx->last_ch = ch;
+    if (ctx->flog) {
+        char tbuf[64];
+        struct timespec ts;
+        clock_gettime(CLOCK_REALTIME, &ts);
+        struct tm *tm_info = gmtime(&ts.tv_sec);
+        size_t len = strftime(tbuf, sizeof(tbuf), "%Y%m%dT%H:%M:%S", tm_info);
+        snprintf(tbuf + len, sizeof(tbuf) - len, ".%06ld", ts.tv_nsec / 1000);
+        fprintf(ctx->flog, "%s Input: %d\\n", tbuf, ch);
+        fflush(ctx->flog);
+    }
+
+    if (ch == 545 || ch == 560 || ch == 443 || ch == 564 || ch == 554) {
+         ctx->procinfoproc->DisplayMode--;
+         if (ctx->procinfoproc->DisplayMode < 1) ctx->procinfoproc->DisplayMode = 6;
+         if (ctx->flog) { fprintf(ctx->flog, "  -> Mode changed to %d\\n", ctx->procinfoproc->DisplayMode); fflush(ctx->flog); }
+    }
+    else if (ch == 561 || ch == 566 || ch == 444 || ch == 565 || ch == 569) {
+         ctx->procinfoproc->DisplayMode++;
+         if (ctx->procinfoproc->DisplayMode > 6) ctx->procinfoproc->DisplayMode = 1;
+         if (ctx->flog) { fprintf(ctx->flog, "  -> Mode changed to %d\\n", ctx->procinfoproc->DisplayMode); fflush(ctx->flog); }
+    }
+    else if (ch >= '0' && ch <= '9') {
+         int cidx = ch - '0';
+         int m = ctx->procinfoproc->DisplayMode;
+         if (m >= 0 && m < 10)
+             ctx->procinfoproc->col_visible[m][cidx] = !ctx->procinfoproc->col_visible[m][cidx];
+    }
+    else if (ch == ANSI_KEY_F2) ctx->procinfoproc->DisplayMode = PROCCTRL_DISPLAYMODE_CTRL;
+    else if (ch == ANSI_KEY_F3) ctx->procinfoproc->DisplayMode = PROCCTRL_DISPLAYMODE_RESOURCES;
+    else if (ch == ANSI_KEY_F4) ctx->procinfoproc->DisplayMode = PROCCTRL_DISPLAYMODE_TRIGGER;
+    else if (ch == ANSI_KEY_F5) ctx->procinfoproc->DisplayMode = PROCCTRL_DISPLAYMODE_TIMING;
+    else if (ch == ANSI_KEY_F6) ctx->procinfoproc->DisplayMode = PROCCTRL_DISPLAYMODE_PROCINFO;
+    else if (ch == 'h')      ctx->procinfoproc->DisplayMode = PROCCTRL_DISPLAYMODE_HELP;
+    else if (ch == ANSI_KEY_CTRL_LEFT) {
+        ctx->procinfoproc->DisplayMode--;
+        if(ctx->procinfoproc->DisplayMode < 1) ctx->procinfoproc->DisplayMode = 6;
+    }
+    else if (ch == ANSI_KEY_CTRL_RIGHT) {
+        ctx->procinfoproc->DisplayMode++;
+        if(ctx->procinfoproc->DisplayMode > 6) ctx->procinfoproc->DisplayMode = 1;
+    }
+    else if (ch == 'x' || ch == 3) { ctx->loopOK = 0; ctx->Xexit = 1; }
+    else if (ch == 'f') { ctx->freeze = !ctx->freeze; }
+    else if (ch == '+' || ch == '=') { ctx->frequ *= 1.2; if (ctx->frequ > 1000.0) ctx->frequ = 1000.0; }
+    else if (ch == '-') { ctx->frequ /= 1.2; if (ctx->frequ < 0.1) ctx->frequ = 0.1; }
+    else if (ch == ' ' && ctx->pindexSelected >= 0) {
+         ctx->procinfoproc->selectedarray[ctx->pindexSelected] = !ctx->procinfoproc->selectedarray[ctx->pindexSelected];
+    }
+    else if (ch == ANSI_KEY_UP && NBactive > 0) {
+         ctx->pindexActiveSelected--; if (ctx->pindexActiveSelected < 0) ctx->pindexActiveSelected = 0;
+    }
+    else if (ch == ANSI_KEY_DOWN && NBactive > 0) {
+         ctx->pindexActiveSelected++; if (ctx->pindexActiveSelected >= NBactive) ctx->pindexActiveSelected = NBactive - 1;
+    }
+    else if (ch == ANSI_KEY_LEFT) {
+         ctx->procinfoproc->selected_col--;
+         if (ctx->procinfoproc->selected_col < 1) ctx->procinfoproc->selected_col = 9;
+    }
+    else if (ch == ANSI_KEY_RIGHT) {
+         ctx->procinfoproc->selected_col++;
+         if (ctx->procinfoproc->selected_col > 9) ctx->procinfoproc->selected_col = 1;
+    }
+    else if (ch == 'r' || ch == 'R') {
+        int all = (ch == 'R');
+        for(int i=0; i<PROCESSINFOLISTSIZE; i++) {
+            if (all || i == ctx->pindexSelected) {
+                if (pinfolist->active[i] == 2 || pinfolist->active[i] == 3) pinfolist->active[i] = 0;
+            }
         }
-
-        if (pindexActiveSelected >= NBactive && NBactive > 0) pindexActiveSelected = NBactive - 1;
-        
-        if (NBactive > 0 && scan_shm != NULL) {
-            if (procinfoproc->sort_col[m] > 0) pindexSelected = procinfoproc->local_sorted_pindex[pindexActiveSelected];
-            else pindexSelected = scan_shm->sorted_pindex[pindexActiveSelected];
+    }
+    else if ((ch == 'T' || ch == 'K' || ch == 'I') && NBactive > 0) {
+        int sig = (ch == 'T') ? SIGTERM : (ch == 'K') ? SIGKILL : SIGINT;
+        int any_sel = 0;
+        for(int i=0; i<PROCESSINFOLISTSIZE; i++) if(ctx->procinfoproc->selectedarray[i]) {
+            kill(pinfolist->PIDarray[i], sig);
+            any_sel = 1;
         }
-        else pindexSelected = -1;
-
-        if (ch != -1) {
-            last_ch = ch;
-            if (flog) {
-                char tbuf[64];
-                struct timespec ts;
-                clock_gettime(CLOCK_REALTIME, &ts);
-                struct tm *tm_info = gmtime(&ts.tv_sec);
-                size_t len = strftime(tbuf, sizeof(tbuf), "%Y%m%dT%H:%M:%S", tm_info);
-                snprintf(tbuf + len, sizeof(tbuf) - len, ".%06ld", ts.tv_nsec / 1000);
-                fprintf(flog, "%s Input: %d\n", tbuf, ch);
-                fflush(flog);
-            }
-
-            // PRIMARY INPUT HANDLER
-            if (ch == 545 || ch == 560 || ch == 443 || ch == 564 || ch == 554) { // CTRL+LEFT
-                 procinfoproc->DisplayMode--;
-                 if (procinfoproc->DisplayMode < 1) procinfoproc->DisplayMode = 6;
-                 if (flog) { fprintf(flog, "  -> Mode changed to %d\n", procinfoproc->DisplayMode); fflush(flog); }
-            }
-            else if (ch == 561 || ch == 566 || ch == 444 || ch == 565 || ch == 569) { // CTRL+RIGHT
-                 procinfoproc->DisplayMode++;
-                 if (procinfoproc->DisplayMode > 6) procinfoproc->DisplayMode = 1;
-                 if (flog) { fprintf(flog, "  -> Mode changed to %d\n", procinfoproc->DisplayMode); fflush(flog); }
-            }
-            else if (ch >= '0' && ch <= '9') {
-                 int cidx = ch - '0';
-                 int m = procinfoproc->DisplayMode;
-                 if (m >= 0 && m < 10)
-                     procinfoproc->col_visible[m][cidx] = !procinfoproc->col_visible[m][cidx];
-            }
-            else if (ch == ANSI_KEY_F2) procinfoproc->DisplayMode = PROCCTRL_DISPLAYMODE_CTRL;
-            else if (ch == ANSI_KEY_F3) procinfoproc->DisplayMode = PROCCTRL_DISPLAYMODE_RESOURCES;
-            else if (ch == ANSI_KEY_F4) procinfoproc->DisplayMode = PROCCTRL_DISPLAYMODE_TRIGGER;
-            else if (ch == ANSI_KEY_F5) procinfoproc->DisplayMode = PROCCTRL_DISPLAYMODE_TIMING;
-            else if (ch == ANSI_KEY_F6) procinfoproc->DisplayMode = PROCCTRL_DISPLAYMODE_PROCINFO;
-            else if (ch == 'h')      procinfoproc->DisplayMode = PROCCTRL_DISPLAYMODE_HELP;
-            else if (ch == ANSI_KEY_CTRL_LEFT) {
-                procinfoproc->DisplayMode--;
-                if(procinfoproc->DisplayMode < 1) procinfoproc->DisplayMode = 6;
-            }
-            else if (ch == ANSI_KEY_CTRL_RIGHT) {
-                procinfoproc->DisplayMode++;
-                if(procinfoproc->DisplayMode > 6) procinfoproc->DisplayMode = 1;
-            }
-            else if (ch == 'x' || ch == 3) { loopOK = 0; Xexit = 1; }
-            else if (ch == 'f') { freeze = !freeze; }
-            else if (ch == '+' || ch == '=') { frequ *= 1.2; if (frequ > 1000.0) frequ = 1000.0; }
-            else if (ch == '-') { frequ /= 1.2; if (frequ < 0.1) frequ = 0.1; }
-            else if (ch == ' ' && pindexSelected >= 0) {
-                 procinfoproc->selectedarray[pindexSelected] = !procinfoproc->selectedarray[pindexSelected];
-            }
-            else if (ch == ANSI_KEY_UP && NBactive > 0) {
-                 pindexActiveSelected--; if (pindexActiveSelected < 0) pindexActiveSelected = 0;
-            }
-            else if (ch == ANSI_KEY_DOWN && NBactive > 0) {
-                 pindexActiveSelected++; if (pindexActiveSelected >= NBactive) pindexActiveSelected = NBactive - 1;
-            }
-            else if (ch == ANSI_KEY_LEFT) {
-                 procinfoproc->selected_col--;
-                 if (procinfoproc->selected_col < 1) procinfoproc->selected_col = 9;
-            }
-            else if (ch == ANSI_KEY_RIGHT) {
-                 procinfoproc->selected_col++;
-                 if (procinfoproc->selected_col > 9) procinfoproc->selected_col = 1;
-            }
-            else if (ch == 'r' || ch == 'R') {
-                int all = (ch == 'R');
-                for(int i=0; i<PROCESSINFOLISTSIZE; i++) {
-                    if (all || i == pindexSelected) {
-                        if (pinfolist->active[i] == 2 || pinfolist->active[i] == 3) pinfolist->active[i] = 0;
-                    }
-                }
-            }
-            else if ((ch == 'T' || ch == 'K' || ch == 'I') && NBactive > 0) {
-                int sig = (ch == 'T') ? SIGTERM : (ch == 'K') ? SIGKILL : SIGINT;
-                int any_sel = 0;
-                for(int i=0; i<PROCESSINFOLISTSIZE; i++) if(procinfoproc->selectedarray[i]) {
-                    kill(pinfolist->PIDarray[i], sig);
-                    any_sel = 1;
-                }
-                if(!any_sel && pindexSelected >= 0) kill(pinfolist->PIDarray[pindexSelected], sig);
-            }
-            else if (ch == 's') {
-                 int m = procinfoproc->DisplayMode;
-                 procinfoproc->sort_mode[m] = m; // Use current mode logic
-                 if (procinfoproc->sort_col[m] == procinfoproc->selected_col) {
-                     procinfoproc->sort_dir[m] = !procinfoproc->sort_dir[m];
-                 } else {
-                     procinfoproc->sort_col[m] = procinfoproc->selected_col;
-                     procinfoproc->sort_dir[m] = 0; // Default ascending
-                 }
-            }
-            else if (ch == 'S') {
-                 int m_curr = procinfoproc->DisplayMode;
-                 int smod = procinfoproc->sort_mode[m_curr];
-                 int scol = procinfoproc->sort_col[m_curr];
-                 int sdir = procinfoproc->sort_dir[m_curr];
-                 for(int m=0; m<10; m++) {
-                     procinfoproc->sort_mode[m] = smod;
-                     procinfoproc->sort_col[m] = scol;
-                     procinfoproc->sort_dir[m] = sdir;
-                 }
-            }
-            else if ((ch == 'p' || ch == 19 || ch == 'e') && NBactive > 0) { // 19 is CTRL+S
-                int val = (ch == 'p') ? -1 : (ch == 19) ? 2 : 3;
-                int any_sel = 0;
-                for(int i=0; i<PROCESSINFOLISTSIZE; i++) if(procinfoproc->selectedarray[i]) {
-                    if (val == -1 && procinfoproc->pinfoarray[i]) procinfoproc->pinfoarray[i]->CTRLval = (procinfoproc->pinfoarray[i]->CTRLval == 0) ? 1 : 0;
-                    else if (procinfoproc->pinfoarray[i]) procinfoproc->pinfoarray[i]->CTRLval = val;
-                    any_sel = 1;
-                }
-                if(!any_sel && pindexSelected >= 0 && procinfoproc->pinfoarray[pindexSelected]) {
-                    if (val == -1) procinfoproc->pinfoarray[pindexSelected]->CTRLval = (procinfoproc->pinfoarray[pindexSelected]->CTRLval == 0) ? 1 : 0;
-                    else procinfoproc->pinfoarray[pindexSelected]->CTRLval = val;
-                }
-            }
-            else if ((ch == 'z' || ch == 'Z') && NBactive > 0) {
-                int all = (ch == 'Z');
-                for(int i=0; i<PROCESSINFOLISTSIZE; i++) if((all || procinfoproc->selectedarray[i]) && procinfoproc->pinfoarray[i]) {
-                    procinfoproc->pinfoarray[i]->loopcnt = 0;
-                }
-                if(!all && !procinfoproc->selectedarray[pindexSelected] && pindexSelected >= 0 && procinfoproc->pinfoarray[pindexSelected]) {
-                    procinfoproc->pinfoarray[pindexSelected]->loopcnt = 0;
-                }
-            }
-            // Removed KEY_RESIZE logic since we clear and resize every frame anyway
+        if(!any_sel && ctx->pindexSelected >= 0) kill(pinfolist->PIDarray[ctx->pindexSelected], sig);
+    }
+    else if (ch == 's') {
+         int m = ctx->procinfoproc->DisplayMode;
+         ctx->procinfoproc->sort_mode[m] = m;
+         if (ctx->procinfoproc->sort_col[m] == ctx->procinfoproc->selected_col) {
+             ctx->procinfoproc->sort_dir[m] = !ctx->procinfoproc->sort_dir[m];
+         } else {
+             ctx->procinfoproc->sort_col[m] = ctx->procinfoproc->selected_col;
+             ctx->procinfoproc->sort_dir[m] = 0;
+         }
+    }
+    else if (ch == 'S') {
+         int m_curr = ctx->procinfoproc->DisplayMode;
+         int smod = ctx->procinfoproc->sort_mode[m_curr];
+         int scol = ctx->procinfoproc->sort_col[m_curr];
+         int sdir = ctx->procinfoproc->sort_dir[m_curr];
+         for(int m=0; m<10; m++) {
+             ctx->procinfoproc->sort_mode[m] = smod;
+             ctx->procinfoproc->sort_col[m] = scol;
+             ctx->procinfoproc->sort_dir[m] = sdir;
+         }
+    }
+    else if ((ch == 'p' || ch == 19 || ch == 'e') && NBactive > 0) {
+        int val = (ch == 'p') ? -1 : (ch == 19) ? 2 : 3;
+        int any_sel = 0;
+        for(int i=0; i<PROCESSINFOLISTSIZE; i++) if(ctx->procinfoproc->selectedarray[i]) {
+            if (val == -1 && ctx->procinfoproc->pinfoarray[i]) ctx->procinfoproc->pinfoarray[i]->CTRLval = (ctx->procinfoproc->pinfoarray[i]->CTRLval == 0) ? 1 : 0;
+            else if (ctx->procinfoproc->pinfoarray[i]) ctx->procinfoproc->pinfoarray[i]->CTRLval = val;
+            any_sel = 1;
         }
+        if(!any_sel && ctx->pindexSelected >= 0 && ctx->procinfoproc->pinfoarray[ctx->pindexSelected]) {
+            if (val == -1) ctx->procinfoproc->pinfoarray[ctx->pindexSelected]->CTRLval = (ctx->procinfoproc->pinfoarray[ctx->pindexSelected]->CTRLval == 0) ? 1 : 0;
+            else ctx->procinfoproc->pinfoarray[ctx->pindexSelected]->CTRLval = val;
+        }
+    }
+    else if ((ch == 'z' || ch == 'Z') && NBactive > 0) {
+        int all = (ch == 'Z');
+        for(int i=0; i<PROCESSINFOLISTSIZE; i++) if((all || ctx->procinfoproc->selectedarray[i]) && ctx->procinfoproc->pinfoarray[i]) {
+            ctx->procinfoproc->pinfoarray[i]->loopcnt = 0;
+        }
+        if(!all && !ctx->procinfoproc->selectedarray[ctx->pindexSelected] && ctx->pindexSelected >= 0 && ctx->procinfoproc->pinfoarray[ctx->pindexSelected]) {
+            ctx->procinfoproc->pinfoarray[ctx->pindexSelected]->loopcnt = 0;
+        }
+    }
+}
 
-        if (freeze == 0) {
+
+
+static void procctrl_render_frame(procctrl_context_t *ctx, int NBactive) {
+    int m = ctx->procinfoproc->DisplayMode;
+    int pindexSelected = ctx->pindexSelected;
+    int pindexActiveSelected = ctx->pindexActiveSelected;
+        if (ctx->freeze == 0) {
             sc_frame_clear();
             TUI_clearscreen(&wrow, &wcol);
-            snprintf(monstring, monstringlen, "Mode %d   PRESS x TO STOP MONITOR", procinfoproc->DisplayMode);
-            TUI_print_header(monstring, '-');
+            snprintf(ctx->monstring, ctx->monstringlen, "Mode %d   PRESS x TO STOP MONITOR", ctx->procinfoproc->DisplayMode);
+            TUI_print_header(ctx->monstring, '-');
             TUI_newline();
             
             {
                 struct stat shm_stat;
                 char shm_list_fname[STRINGMAXLEN_FULLFILENAME];
-                WRITE_FULLFILENAME(shm_list_fname, "%s/processinfo.list.shm", procdname);
+                WRITE_FULLFILENAME(shm_list_fname, "%s/processinfo.list.shm", ctx->procdname);
                 if(stat(shm_list_fname, &shm_stat) == 0) {
                     char timestr[100];
                     struct tm *tm_info = gmtime(&shm_stat.st_mtime);
                     strftime(timestr, 100, "%Y-%m-%d %H:%M:%S", tm_info);
-                    TUI_printfw("List: %-25s (Upd: %s)  ToolCPU: %5.2f%% (%4.1f fps)", shm_list_fname, timestr, tool_cpu_pcnt, actual_fps);
-                    if (flog) TUI_printfw(" K:%d", last_ch);
+                    TUI_printfw("List: %-25s (Upd: %s)  ToolCPU: %5.2f%% (%4.1f fps)", shm_list_fname, timestr, ctx->tool_cpu_pcnt, ctx->actual_fps);
+                    if (ctx->flog) TUI_printfw(" K:%d", ctx->last_ch);
                     TUI_newline();
                 }
             }
             
-            if (procinfoproc->DisplayMode == PROCCTRL_DISPLAYMODE_HELP) screenprint_setcolor(2);
+            if (ctx->procinfoproc->DisplayMode == PROCCTRL_DISPLAYMODE_HELP) screenprint_setcolor(2);
             TUI_printfw("[h] Help");
-            if (procinfoproc->DisplayMode == PROCCTRL_DISPLAYMODE_HELP) screenprint_unsetcolor(2);
+            if (ctx->procinfoproc->DisplayMode == PROCCTRL_DISPLAYMODE_HELP) screenprint_unsetcolor(2);
             TUI_printfw("   ");
 
-            if (procinfoproc->DisplayMode == PROCCTRL_DISPLAYMODE_CTRL) screenprint_setcolor(2);
+            if (ctx->procinfoproc->DisplayMode == PROCCTRL_DISPLAYMODE_CTRL) screenprint_setcolor(2);
             TUI_printfw("[F2] CTRL");
-            if (procinfoproc->DisplayMode == PROCCTRL_DISPLAYMODE_CTRL) screenprint_unsetcolor(2);
+            if (ctx->procinfoproc->DisplayMode == PROCCTRL_DISPLAYMODE_CTRL) screenprint_unsetcolor(2);
             TUI_printfw("   ");
 
-            if (procinfoproc->DisplayMode == PROCCTRL_DISPLAYMODE_RESOURCES) screenprint_setcolor(2);
+            if (ctx->procinfoproc->DisplayMode == PROCCTRL_DISPLAYMODE_RESOURCES) screenprint_setcolor(2);
             TUI_printfw("[F3] Resources");
-            if (procinfoproc->DisplayMode == PROCCTRL_DISPLAYMODE_RESOURCES) screenprint_unsetcolor(2);
+            if (ctx->procinfoproc->DisplayMode == PROCCTRL_DISPLAYMODE_RESOURCES) screenprint_unsetcolor(2);
             TUI_printfw("   ");
 
-            if (procinfoproc->DisplayMode == PROCCTRL_DISPLAYMODE_TRIGGER) screenprint_setcolor(2);
+            if (ctx->procinfoproc->DisplayMode == PROCCTRL_DISPLAYMODE_TRIGGER) screenprint_setcolor(2);
             TUI_printfw("[F4] Triggering");
-            if (procinfoproc->DisplayMode == PROCCTRL_DISPLAYMODE_TRIGGER) screenprint_unsetcolor(2);
+            if (ctx->procinfoproc->DisplayMode == PROCCTRL_DISPLAYMODE_TRIGGER) screenprint_unsetcolor(2);
             TUI_printfw("   ");
 
-            if (procinfoproc->DisplayMode == PROCCTRL_DISPLAYMODE_TIMING) screenprint_setcolor(2);
+            if (ctx->procinfoproc->DisplayMode == PROCCTRL_DISPLAYMODE_TIMING) screenprint_setcolor(2);
             TUI_printfw("[F5] Timing");
-            if (procinfoproc->DisplayMode == PROCCTRL_DISPLAYMODE_TIMING) screenprint_unsetcolor(2);
+            if (ctx->procinfoproc->DisplayMode == PROCCTRL_DISPLAYMODE_TIMING) screenprint_unsetcolor(2);
             TUI_printfw("   ");
 
-            if (procinfoproc->DisplayMode == PROCCTRL_DISPLAYMODE_PROCINFO) screenprint_setcolor(2);
+            if (ctx->procinfoproc->DisplayMode == PROCCTRL_DISPLAYMODE_PROCINFO) screenprint_setcolor(2);
             TUI_printfw("[F6] PROCINFO");
-            if (procinfoproc->DisplayMode == PROCCTRL_DISPLAYMODE_PROCINFO) screenprint_unsetcolor(2);
+            if (ctx->procinfoproc->DisplayMode == PROCCTRL_DISPLAYMODE_PROCINFO) screenprint_unsetcolor(2);
             TUI_newline();
 
             // Column visibility status line
             {
                 const char *colnames[10] = {NULL};
                 int nbcol = 0;
-                switch(procinfoproc->DisplayMode) {
+                switch(ctx->procinfoproc->DisplayMode) {
                     case PROCCTRL_DISPLAYMODE_CTRL:
                         {
                             static const char *names[] = {"", "idx", "status", "pid", "tstart", "pname", "state", "lcnt", "msg", ""};
@@ -572,34 +437,34 @@ errno_t processinfo_CTRLscreen()
                 }
 
                 if (nbcol > 0) {
-                    int m = procinfoproc->DisplayMode;
-                    if (procinfoproc->selected_col > nbcol) procinfoproc->selected_col = nbcol;
-                    if (procinfoproc->selected_col < 1) procinfoproc->selected_col = 1;
+                    int m = ctx->procinfoproc->DisplayMode;
+                    if (ctx->procinfoproc->selected_col > nbcol) ctx->procinfoproc->selected_col = nbcol;
+                    if (ctx->procinfoproc->selected_col < 1) ctx->procinfoproc->selected_col = 1;
 
                     for(int i=1; i<=nbcol; i++) {
                         char colinfo[40];
-                        if (i == procinfoproc->sort_col[m]) {
-                            snprintf(colinfo, 40, "%d:%s%c", i, colnames[i], procinfoproc->sort_dir[m] ? 'v' : '^');
+                        if (i == ctx->procinfoproc->sort_col[m]) {
+                            snprintf(colinfo, 40, "%d:%s%c", i, colnames[i], ctx->procinfoproc->sort_dir[m] ? 'v' : '^');
                         } else {
                             snprintf(colinfo, 40, "%d:%s", i, colnames[i]);
                         }
 
-                        if (i == procinfoproc->selected_col) screenprint_setcolor(10);
-                        if (procinfoproc->col_visible[m][i]) {
+                        if (i == ctx->procinfoproc->selected_col) screenprint_setcolor(10);
+                        if (ctx->procinfoproc->col_visible[m][i]) {
                             screenprint_setbold(); // Bold font, no background highlight
                             TUI_printfw("%s ", colinfo);
                             screenprint_unsetbold();
                         } else {
                             TUI_printfw("%s ", colinfo); // Normal text
                         }
-                        if (i == procinfoproc->selected_col) screenprint_unsetcolor(10);
+                        if (i == ctx->procinfoproc->selected_col) screenprint_unsetcolor(10);
                     }
                     TUI_newline();
                 }
                 TUI_newline();
             }
 
-            if (procinfoproc->DisplayMode == PROCCTRL_DISPLAYMODE_HELP) {
+            if (ctx->procinfoproc->DisplayMode == PROCCTRL_DISPLAYMODE_HELP) {
                 TUI_printfw("MILK Process Control (procCTRL) - HELP");
                 TUI_newline();
                 TUI_printfw("Navigation:  ");
@@ -650,35 +515,35 @@ errno_t processinfo_CTRLscreen()
                     margin_up = 0;
                 }
 
-                while (pindexActiveSelected - doffsetindex > dispindexMax - 1 - margin_dn) {
-                    doffsetindex++;
+                while (ctx->pindexActiveSelected - ctx->doffsetindex > dispindexMax - 1 - margin_dn) {
+                    ctx->doffsetindex++;
                 }
 
-                while (pindexActiveSelected < doffsetindex + margin_up) {
-                    doffsetindex--;
+                while (ctx->pindexActiveSelected < ctx->doffsetindex + margin_up) {
+                    ctx->doffsetindex--;
                 }
 
-                if (pindexActiveSelected < doffsetindex) {
-                    doffsetindex = pindexActiveSelected;
+                if (ctx->pindexActiveSelected < ctx->doffsetindex) {
+                    ctx->doffsetindex = ctx->pindexActiveSelected;
                 }
-                if (pindexActiveSelected >= doffsetindex + dispindexMax) {
-                    doffsetindex = pindexActiveSelected - dispindexMax + 1;
-                }
-
-                if (doffsetindex < 0) {
-                    doffsetindex = 0;
+                if (ctx->pindexActiveSelected >= ctx->doffsetindex + dispindexMax) {
+                    ctx->doffsetindex = ctx->pindexActiveSelected - dispindexMax + 1;
                 }
 
-                int lastindex = doffsetindex + dispindexMax;
+                if (ctx->doffsetindex < 0) {
+                    ctx->doffsetindex = 0;
+                }
+
+                int lastindex = ctx->doffsetindex + dispindexMax;
                 if (lastindex > NBactive) {
                     lastindex = NBactive;
                 }
 
-                for(int dispindex = doffsetindex; dispindex < lastindex; dispindex++) {
-                    if (dispindex == doffsetindex && doffsetindex > 0) {
+                for(int dispindex = ctx->doffsetindex; dispindex < lastindex; dispindex++) {
+                    if (dispindex == ctx->doffsetindex && ctx->doffsetindex > 0) {
                         screenprint_setbold();
                         screenprint_setcolor(3);
-                        TUI_printfw("      ^^^^ %d more entries above ^^^^", (int)(doffsetindex + 1));
+                        TUI_printfw("      ^^^^ %d more entries above ^^^^", (int)(ctx->doffsetindex + 1));
                         screenprint_unsetcolor(3);
                         screenprint_unsetbold();
                         TUI_newline();
@@ -695,28 +560,28 @@ errno_t processinfo_CTRLscreen()
                         continue;
                     }
 
-                if (scan_shm == NULL) break;
+                if (ctx->scan_shm == NULL) break;
                 int pindex;
-                if (procinfoproc->sort_col[m] > 0) pindex = procinfoproc->local_sorted_pindex[dispindex];
-                else pindex = scan_shm->sorted_pindex[dispindex];
+                if (ctx->procinfoproc->sort_col[m] > 0) pindex = ctx->procinfoproc->local_sorted_pindex[dispindex];
+                else pindex = ctx->scan_shm->sorted_pindex[dispindex];
                 
-                if (pindex >= 0 && pindex < PROCESSINFOLISTSIZE && (procinfoproc->pinfommapped[pindex] || pinfolist->active[pindex] != 0)) {
-                        if (pindex == pindexSelected) screenprint_setreverse();
+                if (pindex >= 0 && pindex < PROCESSINFOLISTSIZE && (ctx->procinfoproc->pinfommapped[pindex] || pinfolist->active[pindex] != 0)) {
+                        if (pindex == ctx->pindexSelected) screenprint_setreverse();
                         
-                        if (procinfoproc->selectedarray[pindex]) TUI_printfw("* "); else TUI_printfw("  ");
+                        if (ctx->procinfoproc->selectedarray[pindex]) TUI_printfw("* "); else TUI_printfw("  ");
                         
-                        int m = procinfoproc->DisplayMode;
+                        int m = ctx->procinfoproc->DisplayMode;
 
                         // Column 1: idx
-                        if (procinfoproc->col_visible[m][1]) {
-                            if (procinfoproc->selected_col == 1) screenprint_setcolor(10);
+                        if (ctx->procinfoproc->col_visible[m][1]) {
+                            if (ctx->procinfoproc->selected_col == 1) screenprint_setcolor(10);
                             TUI_printfw("%4d ", pindex);
-                            if (procinfoproc->selected_col == 1) screenprint_unsetcolor(10);
+                            if (ctx->procinfoproc->selected_col == 1) screenprint_unsetcolor(10);
                         }
 
                         // Column 2: status
-                        if (procinfoproc->col_visible[m][2]) {
-                            if (procinfoproc->selected_col == 2) screenprint_setcolor(10);
+                        if (ctx->procinfoproc->col_visible[m][2]) {
+                            if (ctx->procinfoproc->selected_col == 2) screenprint_setcolor(10);
                             if (pinfolist->active[pindex] == 1) {
                                 screenprint_setcolor(6); TUI_printfw("%-10s ", "ACTIVE"); screenprint_unsetcolor(6);
                             } else if (pinfolist->active[pindex] == 2) {
@@ -726,30 +591,30 @@ errno_t processinfo_CTRLscreen()
                             } else {
                                 TUI_printfw("%-10s ", "OFF");
                             }
-                            if (procinfoproc->selected_col == 2) screenprint_unsetcolor(10);
+                            if (ctx->procinfoproc->selected_col == 2) screenprint_unsetcolor(10);
                         }
                         
                         // Column 3: pid
-                        if (procinfoproc->col_visible[m][3]) {
-                            if (procinfoproc->selected_col == 3) screenprint_setcolor(10);
+                        if (ctx->procinfoproc->col_visible[m][3]) {
+                            if (ctx->procinfoproc->selected_col == 3) screenprint_setcolor(10);
                             pid_t pid = pinfolist->PIDarray[pindex];
                             int pid_exists = (kill(pid, 0) == 0);
                             if (!pid_exists) screenprint_setcolor(4);
-                            char state = scan_shm->pinfodisp[pindex].state;
+                            char state = ctx->scan_shm->pinfodisp[pindex].state;
                             if (state == 0) state = ' ';
                             TUI_printfw("%7d %c ", pid, state);
                             if (!pid_exists) screenprint_unsetcolor(4);
-                            if (procinfoproc->selected_col == 3) screenprint_unsetcolor(10);
+                            if (ctx->procinfoproc->selected_col == 3) screenprint_unsetcolor(10);
                         }
 
-                        if (scan_shm && pindex < PROCESSINFOLISTSIZE && pinfolist->active[pindex] == 1 && scan_shm->request_scan[pindex] == 0) {
-                            scan_shm->request_scan[pindex] = 1;
+                        if (ctx->scan_shm && pindex < PROCESSINFOLISTSIZE && pinfolist->active[pindex] == 1 && ctx->scan_shm->request_scan[pindex] == 0) {
+                            ctx->scan_shm->request_scan[pindex] = 1;
                         }
 
                         if (m == PROCCTRL_DISPLAYMODE_CTRL) {
                             // 4: tstart
-                            if (procinfoproc->col_visible[m][4]) {
-                                if (procinfoproc->selected_col == 4) screenprint_setcolor(10);
+                            if (ctx->procinfoproc->col_visible[m][4]) {
+                                if (ctx->procinfoproc->selected_col == 4) screenprint_setcolor(10);
                                 char tbuf[30];
                                 time_t sec = (time_t)pinfolist->createtime[pindex];
                                 long usec = (long)((pinfolist->createtime[pindex] - sec) * 1000000);
@@ -759,150 +624,150 @@ errno_t processinfo_CTRLscreen()
                                          sizeof(tbuf) - 19,
                                          ".%06ld", usec);
                                 TUI_printfw("%s ", tbuf);
-                                if (procinfoproc->selected_col == 4) screenprint_unsetcolor(10);
+                                if (ctx->procinfoproc->selected_col == 4) screenprint_unsetcolor(10);
                             }
 
                             // 5: pname
-                            if (procinfoproc->col_visible[m][5]) {
-                                if (procinfoproc->selected_col == 5) screenprint_setcolor(10);
+                            if (ctx->procinfoproc->col_visible[m][5]) {
+                                if (ctx->procinfoproc->selected_col == 5) screenprint_setcolor(10);
                                 TUI_printfw("%-25s ", pinfolist->pnamearray[pindex]);
-                                if (procinfoproc->selected_col == 5) screenprint_unsetcolor(10);
+                                if (ctx->procinfoproc->selected_col == 5) screenprint_unsetcolor(10);
                             }
 
-                            int   ctrlval = (procinfoproc->pinfoarray[pindex]) ? procinfoproc->pinfoarray[pindex]->CTRLval : 0;
-                            long  loopcnt = scan_shm->pinfodisp[pindex].loopcnt;
-                            char *desc    = scan_shm->pinfodisp[pindex].statusmsg;
+                            int   ctrlval = (ctx->procinfoproc->pinfoarray[pindex]) ? ctx->procinfoproc->pinfoarray[pindex]->CTRLval : 0;
+                            long  loopcnt = ctx->scan_shm->pinfodisp[pindex].loopcnt;
+                            char *desc    = ctx->scan_shm->pinfodisp[pindex].statusmsg;
 
                             // 6: state
-                            if (procinfoproc->col_visible[m][6]) {
-                                if (procinfoproc->selected_col == 6) screenprint_setcolor(10);
+                            if (ctx->procinfoproc->col_visible[m][6]) {
+                                if (ctx->procinfoproc->selected_col == 6) screenprint_setcolor(10);
                                 if (ctrlval == 1) {
                                      screenprint_setcolor(3); screenprint_setblink(); TUI_printfw("C1"); screenprint_unsetcolor(3); screenprint_unsetblink(); TUI_printfw(" ");
                                 } else {
                                      TUI_printfw("C%d ", ctrlval); 
                                 }
-                                if (procinfoproc->selected_col == 6) screenprint_unsetcolor(10);
+                                if (ctx->procinfoproc->selected_col == 6) screenprint_unsetcolor(10);
                             }
 
                             // 7: lcnt
-                            if (procinfoproc->col_visible[m][7]) {
-                                if (procinfoproc->selected_col == 7) screenprint_setcolor(10);
-                                if (loopcnt != procinfoproc->loopcntarray[pindex]) {
+                            if (ctx->procinfoproc->col_visible[m][7]) {
+                                if (ctx->procinfoproc->selected_col == 7) screenprint_setcolor(10);
+                                if (loopcnt != ctx->procinfoproc->loopcntarray[pindex]) {
                                      screenprint_setcolor(6); TUI_printfw("%10ld ", loopcnt); screenprint_unsetcolor(6);
                                 } else {
                                      TUI_printfw("%10ld ", loopcnt); 
                                 }
-                                procinfoproc->loopcntarray[pindex] = loopcnt;
-                                if (procinfoproc->selected_col == 7) screenprint_unsetcolor(10);
+                                ctx->procinfoproc->loopcntarray[pindex] = loopcnt;
+                                if (ctx->procinfoproc->selected_col == 7) screenprint_unsetcolor(10);
                             }
 
                             // 8: msg
-                            if (procinfoproc->col_visible[m][8]) {
-                                if (procinfoproc->selected_col == 8) screenprint_setcolor(10);
+                            if (ctx->procinfoproc->col_visible[m][8]) {
+                                if (ctx->procinfoproc->selected_col == 8) screenprint_setcolor(10);
                                 TUI_printfw("%-30s ", desc); 
-                                if (procinfoproc->selected_col == 8) screenprint_unsetcolor(10);
+                                if (ctx->procinfoproc->selected_col == 8) screenprint_unsetcolor(10);
                             }
                         }
                         else if (m == PROCCTRL_DISPLAYMODE_RESOURCES) {
                             // 4: pname
-                            if (procinfoproc->col_visible[m][4]) {
-                                if (procinfoproc->selected_col == 4) screenprint_setcolor(10);
+                            if (ctx->procinfoproc->col_visible[m][4]) {
+                                if (ctx->procinfoproc->selected_col == 4) screenprint_setcolor(10);
                                 TUI_printfw("%-25s ", pinfolist->pnamearray[pindex]);
-                                if (procinfoproc->selected_col == 4) screenprint_unsetcolor(10);
+                                if (ctx->procinfoproc->selected_col == 4) screenprint_unsetcolor(10);
                             }
                             // 5: prio
-                            if (procinfoproc->col_visible[m][5]) {
-                                if (procinfoproc->selected_col == 5) screenprint_setcolor(10);
-                                TUI_printfw("P%2d ", scan_shm->pinfodisp[pindex].rt_priority); 
-                                if (procinfoproc->selected_col == 5) screenprint_unsetcolor(10);
+                            if (ctx->procinfoproc->col_visible[m][5]) {
+                                if (ctx->procinfoproc->selected_col == 5) screenprint_setcolor(10);
+                                TUI_printfw("P%2d ", ctx->scan_shm->pinfodisp[pindex].rt_priority); 
+                                if (ctx->procinfoproc->selected_col == 5) screenprint_unsetcolor(10);
                             }
                             // 6: cpu
-                            if (procinfoproc->col_visible[m][6]) {
-                                if (procinfoproc->selected_col == 6) screenprint_setcolor(10);
-                                TUI_printfw("CPU:%5.1f%% ", scan_shm->pinfodisp[pindex].subprocCPUloadarray_timeaveraged[0]);
-                                if (procinfoproc->selected_col == 6) screenprint_unsetcolor(10);
+                            if (ctx->procinfoproc->col_visible[m][6]) {
+                                if (ctx->procinfoproc->selected_col == 6) screenprint_setcolor(10);
+                                TUI_printfw("CPU:%5.1f%% ", ctx->scan_shm->pinfodisp[pindex].subprocCPUloadarray_timeaveraged[0]);
+                                if (ctx->procinfoproc->selected_col == 6) screenprint_unsetcolor(10);
                             }
                             // 7: mem
-                            if (procinfoproc->col_visible[m][7]) {
-                                if (procinfoproc->selected_col == 7) screenprint_setcolor(10);
-                                TUI_printfw("MEM:%7ldkB ", scan_shm->pinfodisp[pindex].VmRSSarray[0]);
-                                if (procinfoproc->selected_col == 7) screenprint_unsetcolor(10);
+                            if (ctx->procinfoproc->col_visible[m][7]) {
+                                if (ctx->procinfoproc->selected_col == 7) screenprint_setcolor(10);
+                                TUI_printfw("MEM:%7ldkB ", ctx->scan_shm->pinfodisp[pindex].VmRSSarray[0]);
+                                if (ctx->procinfoproc->selected_col == 7) screenprint_unsetcolor(10);
                             }
                             // 8: thr
-                            if (procinfoproc->col_visible[m][8]) {
-                                if (procinfoproc->selected_col == 8) screenprint_setcolor(10);
-                                TUI_printfw("Thr:%3d ", scan_shm->pinfodisp[pindex].threads);
-                                if (procinfoproc->selected_col == 8) screenprint_unsetcolor(10);
+                            if (ctx->procinfoproc->col_visible[m][8]) {
+                                if (ctx->procinfoproc->selected_col == 8) screenprint_setcolor(10);
+                                TUI_printfw("Thr:%3d ", ctx->scan_shm->pinfodisp[pindex].threads);
+                                if (ctx->procinfoproc->selected_col == 8) screenprint_unsetcolor(10);
                             }
                         }
                         else if (m == PROCCTRL_DISPLAYMODE_TRIGGER) {
                             // 4: pname
-                            if (procinfoproc->col_visible[m][4]) {
-                                if (procinfoproc->selected_col == 4) screenprint_setcolor(10);
+                            if (ctx->procinfoproc->col_visible[m][4]) {
+                                if (ctx->procinfoproc->selected_col == 4) screenprint_setcolor(10);
                                 TUI_printfw("%-25s ", pinfolist->pnamearray[pindex]);
-                                if (procinfoproc->selected_col == 4) screenprint_unsetcolor(10);
+                                if (ctx->procinfoproc->selected_col == 4) screenprint_unsetcolor(10);
                             }
                             // 5: tmode
-                            if (procinfoproc->col_visible[m][5]) {
-                                if (procinfoproc->selected_col == 5) screenprint_setcolor(10);
-                                TUI_printfw("TR:%d ", scan_shm->pinfodisp[pindex].triggermode);
-                                if (procinfoproc->selected_col == 5) screenprint_unsetcolor(10);
+                            if (ctx->procinfoproc->col_visible[m][5]) {
+                                if (ctx->procinfoproc->selected_col == 5) screenprint_setcolor(10);
+                                TUI_printfw("TR:%d ", ctx->scan_shm->pinfodisp[pindex].triggermode);
+                                if (ctx->procinfoproc->selected_col == 5) screenprint_unsetcolor(10);
                             }
                             // 6: tstream
-                            if (procinfoproc->col_visible[m][6]) {
-                                if (procinfoproc->selected_col == 6) screenprint_setcolor(10);
-                                TUI_printfw("%-15s ", scan_shm->pinfodisp[pindex].triggerstreamname);
-                                if (procinfoproc->selected_col == 6) screenprint_unsetcolor(10);
+                            if (ctx->procinfoproc->col_visible[m][6]) {
+                                if (ctx->procinfoproc->selected_col == 6) screenprint_setcolor(10);
+                                TUI_printfw("%-15s ", ctx->scan_shm->pinfodisp[pindex].triggerstreamname);
+                                if (ctx->procinfoproc->selected_col == 6) screenprint_unsetcolor(10);
                             }
                             // 7: tsem
-                            if (procinfoproc->col_visible[m][7]) {
-                                if (procinfoproc->selected_col == 7) screenprint_setcolor(10);
-                                TUI_printfw("S:%d ", scan_shm->pinfodisp[pindex].triggersem);
-                                if (procinfoproc->selected_col == 7) screenprint_unsetcolor(10);
+                            if (ctx->procinfoproc->col_visible[m][7]) {
+                                if (ctx->procinfoproc->selected_col == 7) screenprint_setcolor(10);
+                                TUI_printfw("S:%d ", ctx->scan_shm->pinfodisp[pindex].triggersem);
+                                if (ctx->procinfoproc->selected_col == 7) screenprint_unsetcolor(10);
                             }
                             // 8: tcnt
-                            if (procinfoproc->col_visible[m][8]) {
-                                if (procinfoproc->selected_col == 8) screenprint_setcolor(10);
-                                TUI_printfw("CNT:%ld ", (long)scan_shm->pinfodisp[pindex].triggerstreamcnt);
-                                if (procinfoproc->selected_col == 8) screenprint_unsetcolor(10);
+                            if (ctx->procinfoproc->col_visible[m][8]) {
+                                if (ctx->procinfoproc->selected_col == 8) screenprint_setcolor(10);
+                                TUI_printfw("CNT:%ld ", (long)ctx->scan_shm->pinfodisp[pindex].triggerstreamcnt);
+                                if (ctx->procinfoproc->selected_col == 8) screenprint_unsetcolor(10);
                             }
                             // 9: tmiss
-                            if (procinfoproc->col_visible[m][9]) {
-                                if (procinfoproc->selected_col == 9) screenprint_setcolor(10);
-                                TUI_printfw("M:%d/%ld ", scan_shm->pinfodisp[pindex].triggermissedframe, (long)scan_shm->pinfodisp[pindex].triggermissedframe_cumul);
-                                if (procinfoproc->selected_col == 9) screenprint_unsetcolor(10);
+                            if (ctx->procinfoproc->col_visible[m][9]) {
+                                if (ctx->procinfoproc->selected_col == 9) screenprint_setcolor(10);
+                                TUI_printfw("M:%d/%ld ", ctx->scan_shm->pinfodisp[pindex].triggermissedframe, (long)ctx->scan_shm->pinfodisp[pindex].triggermissedframe_cumul);
+                                if (ctx->procinfoproc->selected_col == 9) screenprint_unsetcolor(10);
                             }
                         }
                         else if (m == PROCCTRL_DISPLAYMODE_TIMING) {
                             // 4: pname
-                            if (procinfoproc->col_visible[m][4]) {
-                                if (procinfoproc->selected_col == 4) screenprint_setcolor(10);
+                            if (ctx->procinfoproc->col_visible[m][4]) {
+                                if (ctx->procinfoproc->selected_col == 4) screenprint_setcolor(10);
                                 TUI_printfw("%-25s ", pinfolist->pnamearray[pindex]);
-                                if (procinfoproc->selected_col == 4) screenprint_unsetcolor(10);
+                                if (ctx->procinfoproc->selected_col == 4) screenprint_unsetcolor(10);
                             }
-                            if (scan_shm->pinfodisp[pindex].MeasureTiming) {
+                            if (ctx->scan_shm->pinfodisp[pindex].MeasureTiming) {
                                 double freq = 0.0;
-                                if (scan_shm->pinfodisp[pindex].dtmedian_iter_ns > 0) freq = 1.0e9 / scan_shm->pinfodisp[pindex].dtmedian_iter_ns;
-                                double exec_ms = 1.0e-6 * scan_shm->pinfodisp[pindex].dtmedian_exec_ns;
+                                if (ctx->scan_shm->pinfodisp[pindex].dtmedian_iter_ns > 0) freq = 1.0e9 / ctx->scan_shm->pinfodisp[pindex].dtmedian_iter_ns;
+                                double exec_ms = 1.0e-6 * ctx->scan_shm->pinfodisp[pindex].dtmedian_exec_ns;
                                 double overhead = 0.0;
-                                if (scan_shm->pinfodisp[pindex].dtmedian_iter_ns > 0) overhead = 100.0 * scan_shm->pinfodisp[pindex].dtmedian_exec_ns / scan_shm->pinfodisp[pindex].dtmedian_iter_ns;
+                                if (ctx->scan_shm->pinfodisp[pindex].dtmedian_iter_ns > 0) overhead = 100.0 * ctx->scan_shm->pinfodisp[pindex].dtmedian_exec_ns / ctx->scan_shm->pinfodisp[pindex].dtmedian_iter_ns;
                                 // 5: freq
-                                if (procinfoproc->col_visible[m][5]) {
-                                    if (procinfoproc->selected_col == 5) screenprint_setcolor(10);
+                                if (ctx->procinfoproc->col_visible[m][5]) {
+                                    if (ctx->procinfoproc->selected_col == 5) screenprint_setcolor(10);
                                     TUI_printfw("%8.2fHz ", freq);
-                                    if (procinfoproc->selected_col == 5) screenprint_unsetcolor(10);
+                                    if (ctx->procinfoproc->selected_col == 5) screenprint_unsetcolor(10);
                                 }
                                 // 6: exec
-                                if (procinfoproc->col_visible[m][6]) {
-                                    if (procinfoproc->selected_col == 6) screenprint_setcolor(10);
+                                if (ctx->procinfoproc->col_visible[m][6]) {
+                                    if (ctx->procinfoproc->selected_col == 6) screenprint_setcolor(10);
                                     TUI_printfw("Exec:%7.3fms ", exec_ms);
-                                    if (procinfoproc->selected_col == 6) screenprint_unsetcolor(10);
+                                    if (ctx->procinfoproc->selected_col == 6) screenprint_unsetcolor(10);
                                 }
                                 // 7: over
-                                if (procinfoproc->col_visible[m][7]) {
-                                    if (procinfoproc->selected_col == 7) screenprint_setcolor(10);
+                                if (ctx->procinfoproc->col_visible[m][7]) {
+                                    if (ctx->procinfoproc->selected_col == 7) screenprint_setcolor(10);
                                     TUI_printfw("(%5.1f%%) ", overhead);
-                                    if (procinfoproc->selected_col == 7) screenprint_unsetcolor(10);
+                                    if (ctx->procinfoproc->selected_col == 7) screenprint_unsetcolor(10);
                                 }
                             } else {
                                 TUI_printfw("--- Timing Disabled ---");
@@ -910,48 +775,48 @@ errno_t processinfo_CTRLscreen()
                         }
                         else if (m == PROCCTRL_DISPLAYMODE_PROCINFO) {
                             // 4: pname
-                            if (procinfoproc->col_visible[m][4]) {
-                                if (procinfoproc->selected_col == 4) screenprint_setcolor(10);
+                            if (ctx->procinfoproc->col_visible[m][4]) {
+                                if (ctx->procinfoproc->selected_col == 4) screenprint_setcolor(10);
                                 TUI_printfw("%-25s ", pinfolist->pnamearray[pindex]);
-                                if (procinfoproc->selected_col == 4) screenprint_unsetcolor(10);
+                                if (ctx->procinfoproc->selected_col == 4) screenprint_unsetcolor(10);
                             }
                             // 5: RT
-                            if (procinfoproc->col_visible[m][5]) {
-                                if (procinfoproc->selected_col == 5) screenprint_setcolor(10);
-                                TUI_printfw("RT:%2d ", scan_shm->pinfodisp[pindex].rt_priority);
-                                if (procinfoproc->selected_col == 5) screenprint_unsetcolor(10);
+                            if (ctx->procinfoproc->col_visible[m][5]) {
+                                if (ctx->procinfoproc->selected_col == 5) screenprint_setcolor(10);
+                                TUI_printfw("RT:%2d ", ctx->scan_shm->pinfodisp[pindex].rt_priority);
+                                if (ctx->procinfoproc->selected_col == 5) screenprint_unsetcolor(10);
                             }
                             // 6: loopMax
-                            if (procinfoproc->col_visible[m][6]) {
-                                if (procinfoproc->selected_col == 6) screenprint_setcolor(10);
-                                TUI_printfw("Lmax:%7ld ", scan_shm->pinfodisp[pindex].loopcntMax);
-                                if (procinfoproc->selected_col == 6) screenprint_unsetcolor(10);
+                            if (ctx->procinfoproc->col_visible[m][6]) {
+                                if (ctx->procinfoproc->selected_col == 6) screenprint_setcolor(10);
+                                TUI_printfw("Lmax:%7ld ", ctx->scan_shm->pinfodisp[pindex].loopcntMax);
+                                if (ctx->procinfoproc->selected_col == 6) screenprint_unsetcolor(10);
                             }
                             // 7: trig
-                            if (procinfoproc->col_visible[m][7]) {
-                                if (procinfoproc->selected_col == 7) screenprint_setcolor(10);
-                                TUI_printfw("Trig:%d ", scan_shm->pinfodisp[pindex].triggermode);
-                                if (procinfoproc->selected_col == 7) screenprint_unsetcolor(10);
+                            if (ctx->procinfoproc->col_visible[m][7]) {
+                                if (ctx->procinfoproc->selected_col == 7) screenprint_setcolor(10);
+                                TUI_printfw("Trig:%d ", ctx->scan_shm->pinfodisp[pindex].triggermode);
+                                if (ctx->procinfoproc->selected_col == 7) screenprint_unsetcolor(10);
                             }
                             // 8: timeout
-                            if (procinfoproc->col_visible[m][8]) {
-                                if (procinfoproc->selected_col == 8) screenprint_setcolor(10);
-                                double tout = scan_shm->pinfodisp[pindex].triggertimeout.tv_sec + 1e-9 * scan_shm->pinfodisp[pindex].triggertimeout.tv_nsec;
+                            if (ctx->procinfoproc->col_visible[m][8]) {
+                                if (ctx->procinfoproc->selected_col == 8) screenprint_setcolor(10);
+                                double tout = ctx->scan_shm->pinfodisp[pindex].triggertimeout.tv_sec + 1e-9 * ctx->scan_shm->pinfodisp[pindex].triggertimeout.tv_nsec;
                                 TUI_printfw("TO:%5.2f ", tout);
-                                if (procinfoproc->selected_col == 8) screenprint_unsetcolor(10);
+                                if (ctx->procinfoproc->selected_col == 8) screenprint_unsetcolor(10);
                             }
                             // 9: timing
-                            if (procinfoproc->col_visible[m][9]) {
-                                if (procinfoproc->selected_col == 9) screenprint_setcolor(10);
-                                TUI_printfw("Tim:%d ", scan_shm->pinfodisp[pindex].MeasureTiming);
-                                if (procinfoproc->selected_col == 9) screenprint_unsetcolor(10);
+                            if (ctx->procinfoproc->col_visible[m][9]) {
+                                if (ctx->procinfoproc->selected_col == 9) screenprint_setcolor(10);
+                                TUI_printfw("Tim:%d ", ctx->scan_shm->pinfodisp[pindex].MeasureTiming);
+                                if (ctx->procinfoproc->selected_col == 9) screenprint_unsetcolor(10);
                             }
                         }
                         else {
                             TUI_printfw("(Mode %d not impl)", m);
                         }
 
-                        if (pindex == pindexSelected) screenprint_unsetreverse();
+                        if (pindex == ctx->pindexSelected) screenprint_unsetreverse();
                         TUI_newline();
                     }
                 }
@@ -959,16 +824,192 @@ errno_t processinfo_CTRLscreen()
             TUI_cleartobottom();
             sc_frame_flush();
         }
+}
+
+
+errno_t processinfo_CTRLscreen()
+{
+    if (getenv("PROCCTRL_DEBUG")) procCTRL_debug_mode = 1;
+    if (procCTRL_debug_mode) printf("DEBUG: processinfo_CTRLscreen start\n");
+
+    procctrl_context_t ctx;
+    memset(&ctx, 0, sizeof(procctrl_context_t));
+
+    if (strlen(procCTRL_logfile) > 0) {
+        ctx.flog = fopen(procCTRL_logfile, "a");
+        if (ctx.flog) {
+            fprintf(ctx.flog, "\n--- processinfo_CTRLscreen started ---\n");
+            fflush(ctx.flog);
+        }
+    }
+
+    if (ctx.flog) { fprintf(ctx.flog, "Checking for daemon...\n"); fflush(ctx.flog); }
+    if (system("pgrep \"milk-procCTRL-s\" > /dev/null") != 0) {
+        fprintf(stderr, "\nWARNING: milk-procCTRL-scan daemon is not running.\n");
+        printf("Start it now in tmux session 'milk-procCTRL-scan'? [y/n] ");
+        fflush(stdout);
+        char response = 'n';
+        if (scanf(" %c", &response) == 1 && (response == 'y' || response == 'Y')) {
+            printf("Launching milk-procCTRL-scan...\n");
+            if(system("tmux new-session -d -s milk-procCTRL-scan 'milk-procCTRL-scan'") < 0) {}
+            sleep(1);
+            if (system("pgrep \"milk-procCTRL-s\" > /dev/null") != 0) {
+                fprintf(stderr, "ERROR: Failed to launch milk-procCTRL-scan daemon.\n");
+                if (ctx.flog) fclose(ctx.flog);
+                return RETURN_FAILURE;
+            }
+        } else {
+            fprintf(stderr, "ERROR: milk-procCTRL-scan daemon is required for this tool.\n");
+            if (ctx.flog) fclose(ctx.flog);
+            return RETURN_FAILURE;
+        }
+    }
+
+    processinfo_procdirname(local_shmdir);
+    processinfo_procdirname(ctx.procdname);
+
+    if (ctx.flog) { fprintf(ctx.flog, "Allocating procinfoproc...\n"); fflush(ctx.flog); }
+    ctx.procinfoproc = (PROCINFOPROC *) calloc(1, sizeof(PROCINFOPROC));
+    if(ctx.procinfoproc == NULL) {
+        PRINT_ERROR("calloc returns NULL pointer");
+        if (ctx.flog) fclose(ctx.flog);
+        return RETURN_FAILURE;
+    }
+
+    ctx.procinfoproc->NBcpus = GetNumberCPUs(ctx.procinfoproc);
+    GetCPUloads(ctx.procinfoproc); 
+
+    if(system("which cset > /dev/null 2>&1") == 0) ctx.procinfoproc->has_cset = 1;
+    else ctx.procinfoproc->has_cset = 0;
+
+    for(int m=0; m<10; m++) {
+        for(int i=0; i<10; i++) ctx.procinfoproc->col_visible[m][i] = 1;
+        ctx.procinfoproc->sort_col[m] = 0;
+        ctx.procinfoproc->sort_dir[m] = 0;
+        ctx.procinfoproc->sort_mode[m] = m;
+    }
+    ctx.procinfoproc->selected_col = 1;
+
+    STRINGLISTENTRY *CPUsetList = (STRINGLISTENTRY *) malloc(sizeof(STRINGLISTENTRY) * 1000);
+    int NBCPUset __attribute__((unused)) = processinfo_CPUsets_List(CPUsetList, ctx.procinfoproc->has_cset);
+
+    if (ctx.flog) { fprintf(ctx.flog, "Connecting to process list...\n"); fflush(ctx.flog); }
+    if(processinfo_shm_list_create() == -1) {
+        printf("==== ERROR: CANNOT ACCESS PROCESS LIST ====\n");
+        if (ctx.flog) fclose(ctx.flog);
+        return RETURN_FAILURE;
+    }
+    ctx.procinfoproc->pinfolist = pinfolist;
+    
+    char scan_shm_name[STRINGMAXLEN_FULLFILENAME];
+    snprintf(scan_shm_name, sizeof(scan_shm_name), "%s/%s", ctx.procdname, PROCESSINFO_SCAN_SHM_NAME);
+    if (ctx.flog) { fprintf(ctx.flog, "Linking to scan SHM: %s\n", scan_shm_name); fflush(ctx.flog); }
+    ctx.scan_shm = (PROCSCAN_SHM *) link_scan_shm(scan_shm_name, sizeof(PROCSCAN_SHM));
+    if (ctx.scan_shm == MAP_FAILED) {
+        printf("WARNING: Could not link to scan SHM %s. Stats may be missing.\n", scan_shm_name);
+        ctx.scan_shm = NULL;
+    }
+
+    TUI_set_screenprintmode(SCREENPRINT_NCURSES);
+    if(getenv("MILK_TUIPRINT_STDIO")) TUI_set_screenprintmode(SCREENPRINT_STDIO);
+    
+    if (ctx.flog) { fprintf(ctx.flog, "Initializing terminal...\n"); fflush(ctx.flog); }
+    ansi_raw_mode_enter();
+    TUI_init_terminal(&wrow, &wcol);
+    if(wrow < 10) wrow = 10;
+
+    if (ctx.flog) { fprintf(ctx.flog, "Allocating pinfodisp buffer (250MB)\n"); fflush(ctx.flog); }
+    ctx.procinfoproc->pinfodisp = (PROCESSINFODISP *) calloc(PROCESSINFOLISTSIZE, sizeof(PROCESSINFODISP));
+    if (ctx.procinfoproc->pinfodisp == NULL) {
+        ansi_raw_mode_exit();
+        fprintf(stderr, "FATAL ERROR: Could not allocate 250MB process info buffer.\n");
+        if (ctx.flog) fclose(ctx.flog);
+        return RETURN_FAILURE;
+    }
+    ctx.procinfoproc->NBpinfodisp = PROCESSINFOLISTSIZE;
+    
+    ctx.procinfoproc->DisplayMode = PROCCTRL_DISPLAYMODE_CTRL;
+    ctx.procinfoproc->loop = 1;
+    ctx.procinfoproc->twaitus = 1000000;
+
+    int backstderr = -1, newstderr = -1;
+    if (procCTRL_debug_mode == 0) {
+        fflush(stderr);
+        backstderr = dup(STDERR_FILENO);
+        newstderr  = open("/dev/null", O_WRONLY);
+        dup2(newstderr, STDERR_FILENO);
+        close(newstderr);
+    }
+
+    ctx.t_last_scan.tv_sec = 0; ctx.t_last_scan.tv_nsec = 0;
+    ctx.frequ = 32.0;
+    ctx.pindexSelected = -1;
+    ctx.pindexActiveSelected = 0;
+    ctx.doffsetindex = 0;
+    ctx.freeze = 0;
+    ctx.loopOK = 1;
+    ctx.Xexit = 0;
+    ctx.monstringlen = 200;
+    ctx.last_ch = -1;
+
+    getrusage(RUSAGE_SELF, &ctx.usage_prev);
+    clock_gettime(CLOCK_MONOTONIC, &ctx.t_usage_prev);
+    ctx.tool_cpu_pcnt = 0.0;
+
+    clock_gettime(CLOCK_MONOTONIC, &ctx.t_disp_prev);
+    ctx.actual_fps = 0.0;
+
+    if (ctx.flog) { fprintf(ctx.flog, "Entering main loop.\n"); fflush(ctx.flog); }
+    sc_frame_clear();
+
+    while(ctx.loopOK) {
+        if(processinfo_signal_SEGV) { 
+            if (ctx.flog) { fprintf(ctx.flog, "SEGV signal received!\n"); fflush(ctx.flog); }
+            ctx.loopOK=0; break; 
+        }
+
+        procctrl_update_stats(&ctx);
+
+        usleep((long)(1000000.0 / ctx.frequ));
+        int ch = get_singlechar_nonblock();
+        
+        int NBactive = (ctx.scan_shm) ? ctx.scan_shm->NBactive : 0;
+        int m = ctx.procinfoproc->DisplayMode;
+        
+        if (ctx.procinfoproc->sort_col[m] > 0 && ctx.scan_shm != NULL) {
+            sort_ctx_m = ctx.procinfoproc->sort_mode[m];
+            sort_ctx_col = ctx.procinfoproc->sort_col[m];
+            sort_ctx_dir = ctx.procinfoproc->sort_dir[m];
+            sort_ctx_scan_shm = ctx.scan_shm;
+            sort_ctx_pinfolist = pinfolist;
+            
+            for(int i=0; i<NBactive; i++) ctx.procinfoproc->local_sorted_pindex[i] = ctx.scan_shm->sorted_pindex[i];
+            qsort(ctx.procinfoproc->local_sorted_pindex, NBactive, sizeof(int), proc_comp);
+        }
+
+        if (ctx.pindexActiveSelected >= NBactive && NBactive > 0) ctx.pindexActiveSelected = NBactive - 1;
+        
+        if (NBactive > 0 && ctx.scan_shm != NULL) {
+            if (ctx.procinfoproc->sort_col[m] > 0) ctx.pindexSelected = ctx.procinfoproc->local_sorted_pindex[ctx.pindexActiveSelected];
+            else ctx.pindexSelected = ctx.scan_shm->sorted_pindex[ctx.pindexActiveSelected];
+        }
+        else ctx.pindexSelected = -1;
+
+        if (ch != -1) {
+            procctrl_handle_keyboard_event(&ctx, ch, NBactive);
+        }
+
+        procctrl_render_frame(&ctx, NBactive);
     }
 
     ansi_raw_mode_exit();
-    if (scan_shm) munmap(scan_shm, sizeof(PROCSCAN_SHM));
-    for(long i=0; i<PROCESSINFOLISTSIZE; i++) if(procinfoproc->pinfommapped[i]) {
-        if (procinfoproc->pinfoarray[i] != NULL && procinfoproc->pinfoarray[i] != (PROCESSINFO*)MAP_FAILED)
-            processinfo_shm_close(procinfoproc->pinfoarray[i], procinfoproc->fdarray[i]);
+    if (ctx.scan_shm) munmap(ctx.scan_shm, sizeof(PROCSCAN_SHM));
+    for(long i=0; i<PROCESSINFOLISTSIZE; i++) if(ctx.procinfoproc->pinfommapped[i]) {
+        if (ctx.procinfoproc->pinfoarray[i] != NULL && ctx.procinfoproc->pinfoarray[i] != (PROCESSINFO*)MAP_FAILED)
+            processinfo_shm_close(ctx.procinfoproc->pinfoarray[i], ctx.procinfoproc->fdarray[i]);
     }
-    free(procinfoproc->pinfodisp);
-    free(procinfoproc);
+    free(ctx.procinfoproc->pinfodisp);
+    free(ctx.procinfoproc);
     free(CPUsetList);
 
     if (procCTRL_debug_mode == 0) {
@@ -976,6 +1017,6 @@ errno_t processinfo_CTRLscreen()
         dup2(backstderr, STDERR_FILENO);
         close(backstderr);
     }
-    if (flog) { fprintf(flog, "--- processinfo_CTRLscreen ended ---\n"); fclose(flog); }
+    if (ctx.flog) { fprintf(ctx.flog, "--- processinfo_CTRLscreen ended ---\n"); fclose(ctx.flog); }
     return RETURN_SUCCESS;
 }
